+++ tests/Replication2/CMakeLists.txt
@@ -31,10 +31,14 @@ set(ARANGODB_REPLICATION2_TEST_HELPER_SOURCES
 
 set(ARANGODB_REPLICATION2_TEST_SOURCES
   ReplicatedLog/Integration/MaintenanceTests.cpp
-  ReplicatedState/DocumentStateMachineTest.cpp
-  ReplicatedState/DocumentStateShardHandlerTest.cpp
   ReplicatedLog/Integration/StorageEngineMethodsTest.cpp
   ReplicatedState/StateManagerTest.cpp
+  ReplicatedState/StateMachines/DocumentState/CoreTest.cpp
+  ReplicatedState/StateMachines/DocumentState/FollowerTest.cpp
+  ReplicatedState/StateMachines/DocumentState/LeaderTest.cpp
+  ReplicatedState/StateMachines/DocumentState/ShardHandlerTest.cpp
+  ReplicatedState/StateMachines/DocumentState/SnapshotTest.cpp
+  ReplicatedState/StateMachines/DocumentState/TransactionHandlerTest.cpp
 
   CollectionGroups/CollectionGroupsSupervision.cpp)
